Update nba.py
I updated the code to use the full nbaRaw.csv dataset and included a little section at the bottom to calculate accuracy based on the distance between predicted playoff result and true playoff result. 

So far the basic accuracy is consistently 60% and the distance accuracy is 87%

We still need to normalize the data by year compared to the league average and account for the many different missing data sections.